requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Deprecated==1.2.7
future==0.18.2
fuzzywuzzy==0.17.0
idna==2.8
joblib==0.14.1
nltk==3.5
openreview-py==1.0.9
ortools==7.4.7247
phrasemachine==1.0.7
pptree==2.0
protobuf==3.11.2
pycryptodome==3.9.4
pylatexenc==2.1
regex==2020.4.4
requests==2.22.0
six==1.14.0
tld==0.10
tqdm==4.42.0
urllib3==1.25.8
wrapt==1.11.2
